+++ lib/rest/autopilot/v1/assistant/query.d.ts
@@ -174,6 +174,7 @@ interface QueryListInstanceCreateOptions {
  * @property callback -
  *                         Function to process each record. If this and a positional
  *                         callback are passed, this one will be used
+ * @property dialogueSid - The SID of the [Dialogue](https://www.twilio.com/docs/autopilot/api/dialogue).
  * @property done - Function to be called upon completion of streaming
  * @property language - The ISO language-country string that specifies the language used by the Query resources to read
  * @property limit -
@@ -191,6 +192,7 @@ interface QueryListInstanceCreateOptions {
  */
 interface QueryListInstanceEachOptions {
   callback?: (item: QueryInstance, done: (err?: Error) => void) => void;
+  dialogueSid?: string;
   done?: Function;
   language?: string;
   limit?: number;
@@ -202,6 +204,7 @@ interface QueryListInstanceEachOptions {
 /**
  * Options to pass to list
  *
+ * @property dialogueSid - The SID of the [Dialogue](https://www.twilio.com/docs/autopilot/api/dialogue).
  * @property language - The ISO language-country string that specifies the language used by the Query resources to read
  * @property limit -
  *                         Upper limit for the number of records to return.
@@ -217,6 +220,7 @@ interface QueryListInstanceEachOptions {
  * @property status - The status of the resources to read
  */
 interface QueryListInstanceOptions {
+  dialogueSid?: string;
   language?: string;
   limit?: number;
   modelBuild?: string;
@@ -227,6 +231,7 @@ interface QueryListInstanceOptions {
 /**
  * Options to pass to page
  *
+ * @property dialogueSid - The SID of the [Dialogue](https://www.twilio.com/docs/autopilot/api/dialogue).
  * @property language - The ISO language-country string that specifies the language used by the Query resources to read
  * @property modelBuild - The SID or unique name of the Model Build to be queried
  * @property pageNumber - Page Number, this value is simply for client state
@@ -235,6 +240,7 @@ interface QueryListInstanceOptions {
  * @property status - The status of the resources to read
  */
 interface QueryListInstancePageOptions {
+  dialogueSid?: string;
   language?: string;
   modelBuild?: string;
   pageNumber?: number;