+++ spec/integration/rest/api/v2010/account/conference/participant.spec.js
@@ -59,6 +59,39 @@ describe('Participant', function() {
       var body = {
           'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': null,
+          'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
+          'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
+          'end_conference_on_exit': false,
+          'muted': false,
+          'hold': false,
+          'status': 'complete',
+          'start_conference_on_enter': true,
+          'coaching': true,
+          'call_sid_to_coach': 'CAbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
+          'uri': '/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conferences/CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants/CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.json'
+      };
+
+      holodeck.mock(new Response(200, body));
+
+      var promise = client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+                                    .conferences('CFXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+                                    .participants('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX').fetch();
+      promise.then(function(response) {
+        expect(response).toBeDefined();
+        done();
+      }, function() {
+        throw new Error('failed');
+      }).done();
+    }
+  );
+  it('should generate valid fetch_by_label response',
+    function(done) {
+      var body = {
+          'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': 'customer',
           'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
           'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -115,6 +148,39 @@ describe('Participant', function() {
       var body = {
           'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': null,
+          'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
+          'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
+          'end_conference_on_exit': false,
+          'muted': true,
+          'hold': false,
+          'status': 'complete',
+          'start_conference_on_enter': true,
+          'coaching': false,
+          'call_sid_to_coach': null,
+          'uri': '/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conferences/CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants/CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.json'
+      };
+
+      holodeck.mock(new Response(200, body));
+
+      var promise = client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+                                    .conferences('CFXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+                                    .participants('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX').update();
+      promise.then(function(response) {
+        expect(response).toBeDefined();
+        done();
+      }, function() {
+        throw new Error('failed');
+      }).done();
+    }
+  );
+  it('should generate valid mute_participant_by_label response',
+    function(done) {
+      var body = {
+          'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': 'customer',
           'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
           'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -146,6 +212,39 @@ describe('Participant', function() {
       var body = {
           'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': null,
+          'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
+          'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
+          'end_conference_on_exit': false,
+          'muted': false,
+          'hold': false,
+          'status': 'complete',
+          'start_conference_on_enter': true,
+          'coaching': true,
+          'call_sid_to_coach': 'CAbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
+          'uri': '/2010-04-01/Accounts/ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Conferences/CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Participants/CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.json'
+      };
+
+      holodeck.mock(new Response(200, body));
+
+      var promise = client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+                                    .conferences('CFXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+                                    .participants('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX').update();
+      promise.then(function(response) {
+        expect(response).toBeDefined();
+        done();
+      }, function() {
+        throw new Error('failed');
+      }).done();
+    }
+  );
+  it('should generate valid modify_participant_by_label response',
+    function(done) {
+      var body = {
+          'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': 'customer',
           'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
           'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -204,6 +303,7 @@ describe('Participant', function() {
       var body = {
           'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': 'customer',
           'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
           'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -236,6 +336,7 @@ describe('Participant', function() {
       var body = {
           'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': 'customer',
           'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
           'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -268,6 +369,7 @@ describe('Participant', function() {
       var body = {
           'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': null,
           'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
           'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -300,6 +402,7 @@ describe('Participant', function() {
       var body = {
           'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': null,
           'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
           'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -332,6 +435,7 @@ describe('Participant', function() {
       var body = {
           'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'label': null,
           'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
           'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -401,6 +505,23 @@ describe('Participant', function() {
       }).done();
     }
   );
+  it('should generate valid delete_by_label response',
+    function(done) {
+      var body = null;
+
+      holodeck.mock(new Response(204, body));
+
+      var promise = client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+                                    .conferences('CFXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+                                    .participants('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX').remove();
+      promise.then(function(response) {
+        expect(response).toBe(true);
+        done();
+      }, function() {
+        throw new Error('failed');
+      }).done();
+    }
+  );
   it('should treat the first each arg as a callback',
     function(done) {
       var body = {
@@ -408,6 +529,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Sat, 19 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Sat, 19 Feb 2011 21:07:19 +0000',
@@ -423,6 +545,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -458,6 +581,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Sat, 19 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Sat, 19 Feb 2011 21:07:19 +0000',
@@ -473,6 +597,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -513,6 +638,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Sat, 19 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Sat, 19 Feb 2011 21:07:19 +0000',
@@ -528,6 +654,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -614,6 +741,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Sat, 19 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Sat, 19 Feb 2011 21:07:19 +0000',
@@ -629,6 +757,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',
@@ -672,6 +801,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAcccccccccccccccccccccccccccccccc',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Thu, 17 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Thu, 17 Feb 2011 21:07:19 +0000',
@@ -687,6 +817,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAdddddddddddddddddddddddddddddddd',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Wed, 16 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Wed, 16 Feb 2011 21:07:19 +0000',
@@ -730,6 +861,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Sat, 19 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Sat, 19 Feb 2011 21:07:19 +0000',
@@ -745,6 +877,7 @@ describe('Participant', function() {
               {
                   'account_sid': 'ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'call_sid': 'CAbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
+                  'label': null,
                   'conference_sid': 'CFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'date_created': 'Fri, 18 Feb 2011 21:07:19 +0000',
                   'date_updated': 'Fri, 18 Feb 2011 21:07:19 +0000',