+++ spec/integration/rest/preview/trusted_comms/business/brand/brandedChannel/channel.spec.js
@@ -33,7 +33,7 @@ describe('Channel', function() {
     function(done) {
       holodeck.mock(new Response(500, {}));
 
-      var opts = {phoneNumberSid: 'PNXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX', phoneNumber: 'phone_number'};
+      var opts = {phoneNumberSid: 'PNXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX'};
       var promise = client.preview.trusted_comms.businesses('BXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
                                                 .brands('BZXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
                                                 .brandedChannels('BWXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
@@ -50,7 +50,7 @@ describe('Channel', function() {
       var brandedChannelSid = 'BWXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX';
       var url = `https://preview.twilio.com/TrustedComms/Businesses/${businessSid}/Brands/${brandSid}/BrandedChannels/${brandedChannelSid}/Channels`;
 
-      var values = {PhoneNumberSid: 'PNXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX', PhoneNumber: 'phone_number', };
+      var values = {PhoneNumberSid: 'PNXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX', };
       holodeck.assertHasRequest(new Request({
           method: 'POST',
           url: url,
@@ -72,7 +72,7 @@ describe('Channel', function() {
 
       holodeck.mock(new Response(201, body));
 
-      var opts = {phoneNumberSid: 'PNXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX', phoneNumber: 'phone_number'};
+      var opts = {phoneNumberSid: 'PNXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX'};
       var promise = client.preview.trusted_comms.businesses('BXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
                                                 .brands('BZXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
                                                 .brandedChannels('BWXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')